virtio-pci: Add Device Memory Buffer (DMB) support
In some environments, especially in confidential computing, DMA into
guest memory is not always desireable or possible. To still allow full
virtio communication, add a new feature bit and exposed shmem region to
virtio which limits all (non-MSI-X) DMA to this prepopulated shmem
region for a given device.

This way, the hypervisor can set up a shared memory region in the BAR
which the guest can access as always unencrypted memory. Compared to
alternatives such as swiotlb, this mechanism reduces noisy neighbor
effects of transactions and does not rely on hypervisor APIs to modify
visibility of a given memory range.

Implement support for DMB in QEMU's virtio-pci layer. Add a new "dmb-size"
property to virtio-pci devices which users can set to enable DMB. When
the size is non-0, QEMU exposes the DMB feature bit and a DMB shmem region
which the guest can then use to perform DMA operations on.

Alternatives considered:

  - https://lore.kernel.org/all/20250402112410.2086892-1-dwmw2@infradead.org/
    - Bleeds abstraction beyond PCI device. This prevents use cases such as
      real hardware implementing this feature.
  - virtio-msg. It can address the same use case, but requires dramatically
    different transport logic, which requires a lot more changes in both
    hypervisor and guest.

Signed-off-by: Alexander Graf <graf@amazon.com>